+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - added extra line if a footer exists to be placed before the footer (so content doesn't mush up)
 - added redirect functionality - make the first page of the content 30 or 31 followed by the url to redirect to
+- some more descriptive startup messages
 ### Fixed
 
 ## [0.2.4]