+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unrelease
 
-## [Unreleased]
+### Added
+### Fixed
+
+## [0.2.6]
 
 ### Added
 - added extra line if a footer exists to be placed before the footer (so content doesn't mush up)